+++ src/canvas.js
@@ -1,3 +1,14 @@
+/**
+ * =============================================================================
+ * Functions:
+ * =============================================================================
+ * Exported:
+ * ---------------------------------------------------------------------------- 
+ * chartData(theData, ...element)
+ * 
+ * chartDataYear(theData, element)
+ * =============================================================================
+*/
 export const chartData = (theData, ...element) => {
   const labels = Object.keys(theData);
   const values = Object.values(theData);
