+++ src/main.js
@@ -1,10 +1,9 @@
 import {dataJson, filter, sort, search, canvas, canvasYear} from './data.js';
 import {chartData, chartDataYear} from './canvas.js';
-
+import {printData, createPaginator, showTable, showCards} from './show.js';
 
 const dir = './data/countries/countries.json';
 const lines = 10;
-const titles = ['No', 'Country', 'Capital', 'Languages', 'Area', 'Population', 'Gini'];
 const filterOptions = ['Continents', 'Subregion', 'Languages'];
 const subFilterOptions = [[],[],[]];
 const arrayOfYears = [];
@@ -20,6 +19,16 @@ const backBut = document.querySelector('#back-button');
 const forwardBut = document.querySelector('#forward-button');
 const table = document.querySelector('table');
 const cards = document.querySelector('#cards');
+const navBarDataButton = document.querySelector('#data');
+const navBarCalculusButton = document.querySelector('#calculus');
+const navBarMapButton = document.querySelector('#map');
+const dataContainer = document.querySelector('#data-container');
+const calculusContainer = document.querySelector('#calculus-container');
+const mapContainer = document.querySelector('#map-container');
+const botonesPaginatorCalculus = document.querySelectorAll('section[id="calculus-container"] nav button');
+const containerGiniGraphYear = document.querySelector('section[data-test="gini-canvas-year"]');
+const containerGiniGraph = document.querySelector('section[data-test="gini-canvas"]');
+const containerClock = document.querySelector('#clockdate');
 
 let yearSelector;
 let totalPages;
@@ -39,24 +48,23 @@ function init() {
   startTime();
   totalPages = Math.ceil(globalData.length/lines);
   document.querySelector('article h3').innerHTML = "Flags of Countries";
-  inputSearch.value = '';
+  //inputSearch.value = '';
   //filterBut.disabled = true;
+  inputToggle.checked = false;
   backBut.disabled = true;
-  createPaginator(totalPages);
+  createPaginator(totalPages, pageSelector);
   createSelectorForFilterBy(); //Create the select for filter by
   theData = globalData;
-  showCards(1, theData);
-  showTable(1, theData);
+  showCards(theData, cards, parseInt(pageSelector.value) + 1, lines);
+  showTable(theData, table, parseInt(pageSelector.value) + 1, lines);
   (inputToggle.checked === true) ? cards.style.display = 'none' : table.style.display = 'none';
-  //createSelectForYears();
-  // const dataGiniYears = canvasYear(globalData, arrayOfYears[yearSelector.value]);
-  // const containerGiniGraphYears = document.querySelector('#gini-canvas-year');
-  // chartDataYear(dataGiniYears, containerGiniGraphYears);
-
-  //const dataGini = canvas(globalData, arrayOfYears[yearSelector.value]);
-  //const containerGiniGraph = document.querySelector('#gini-canvas');
-  //chartData(dataGini, containerGiniGraph, searchData());  
-
+  dataContainer.style.display = "block";
+  calculusContainer.style.display = "none";
+  mapContainer.style.display = "none";
+  
+  const paginationNumbers = document.getElementById("pagination-numbers");
+  const nextButton = document.getElementById("next-button");
+  const prevButton = document.getElementById("prev-button");
   createEventListeners();
 }
 
@@ -94,18 +102,97 @@ function createEventListeners(){
   //yearSelector.addEventListener('change', () => {
   //  graphGiniIndex();
   //});
+  navBarDataButton.addEventListener('click', () => {
+    document.querySelector('article h3').innerHTML = "Flags of Countries";
+    dataContainer.style.display = "block";
+    calculusContainer.style.display = "none";
+    mapContainer.style.display = "none";  
+  });
+  navBarCalculusButton.addEventListener('click', () => {
+    document.querySelector('article h3').innerHTML = "Calculus";
+    dataContainer.style.display = "none";
+    calculusContainer.style.display = "flex";
+    containerGiniGraphYear.style.display = "block";
+    containerGiniGraph.style.display = "none";
+    containerClock.style.display = "none";
+    mapContainer.style.display = "none";
+    createSelectForYears();
+    graphGiniIndex(0);
+    yearSelector.addEventListener('change', ()=>{
+      graphGiniIndex(0);
+    });
+
+  });
+  navBarMapButton.addEventListener('click', () => {
+    dataContainer.style.display = "none";
+    calculusContainer.style.display = "none";
+    mapContainer.style.display = "flex";  
+  });
+
+
+  console.log(botonesPaginatorCalculus);
+  botonesPaginatorCalculus.forEach(element => element.addEventListener('click', (event) =>{
+    console.log(event.currentTarget);
+    event.target.classList.add("active");
+    console.log(event.target.classList);
+    disableButton(event.target);
+    botonesPaginatorCalculus.forEach(element => {
+      if (element.id !== event.target.id){
+        enableButton(element);
+      }
+      if (event.target.id === 'but-4'){
+        disableButton(botonesPaginatorCalculus[botonesPaginatorCalculus.length - 1]);
+      }
+      if (event.target.id === 'but-1'){
+        disableButton(botonesPaginatorCalculus[0]);
+      }
+    });
+    if(event.target.id === 'but-1'){
+      containerGiniGraphYear.style.display = "block";
+      containerGiniGraph.style.display = "none";
+      containerClock.style.display = "none";
+      createSelectForYears();
+      //containerGiniGraph.append(yearSelector);
+      graphGiniIndex(0);    
+    } else if (event.target.id === 'but-2'){
+      containerGiniGraphYear.style.display = "none";
+      containerGiniGraph.style.display = "block";
+      containerClock.style.display = "none";
+      graphGiniIndex('Colombia', 1);
+    } else if (event.target.id === 'but-3'){
+      containerGiniGraphYear.style.display = "none";
+      containerGiniGraph.style.display = "none";
+      containerClock.style.display = "flex";
+    } else if (event.target.id === 'but-4'){
+      
+    }
+  }));
+}
+function disableButton(button){
+  button.classList.add("disabled");
+  button.setAttribute("disabled", true);
+}
+
+function enableButton(button){
+  button.classList.remove("disabled");
+  button.removeAttribute("disabled");
 }
 
 function graphGiniIndex(...extra){
-  // const dataGiniYears = canvasYear(globalData, arrayOfYears[yearSelector.value]);
-  // const canvasGiniGraphYears = document.querySelector('#gini-canvas-year');  
-  // chartDataYear(dataGiniYears, canvasGiniGraphYears);
-  if(extra){
-    alert("Extra");
+  if(parseInt(extra[extra.length-1]) === 0){
+    const dataGiniYears = canvasYear(globalData, arrayOfYears[yearSelector.value]);
+    const canvasGiniGraphYears = document.querySelector('#gini-canvas-year');  
+    chartDataYear(dataGiniYears, canvasGiniGraphYears);
+  } else if(parseInt(extra[extra.length-1]) === 1){
+    const dataGini = canvas(globalData);
+    const containerGiniGraph = document.querySelector('#gini-canvas');
+    chartData(dataGini, containerGiniGraph, 'Colombia');
   }
-  const dataGini = canvas(globalData, arrayOfYears[yearSelector.value]);
-  const containerGiniGraph = document.querySelector('#gini-canvas');
-  chartData(dataGini, containerGiniGraph, searchData());
+  if(extra.length > 1){
+    const dataGini = canvas(globalData, arrayOfYears[yearSelector.value]);
+    const containerGiniGraph = document.querySelector('#gini-canvas');
+    chartData(dataGini, containerGiniGraph, extra[0]);
+  }
 }
 
 function toggleView() {
@@ -118,7 +205,7 @@ function toggleView() {
     table.style.display = 'none';
     document.querySelector('article h3').innerHTML = "Flags of Countries";
   }
-  printData();
+  printData(theData, cards, table, backBut, forwardBut, pageSelector, lines, inputToggle.checked);
 }
 
 function searchData(){
@@ -128,7 +215,7 @@ function searchData(){
     const optionFilterBy = subFilterOptions[selectFilter.value][selectSubFilter.value];
     theData = filter(theData, filterBy, optionFilterBy);
   }
-  printData();
+  printData(theData, cards, table, backBut, forwardBut, pageSelector, lines, inputToggle.checked);
 }
 
 function filterData(){
@@ -142,7 +229,7 @@ function filterData(){
     } else {
       theData = globalData;
     }
-    printData();
+    printData(theData, cards, table, backBut, forwardBut, pageSelector, lines, inputToggle.checked);
   }
 
   //Filtering with the option of the subfilter
@@ -172,7 +259,7 @@ function subFilter(){
   if (inputSearch.value !== ''){
     theData = search(theData, inputSearch.value);
   }
-  printData();
+  printData(theData, cards, table, backBut, forwardBut, pageSelector, lines, inputToggle.checked);
 }
 
 function sortData(event){
@@ -181,7 +268,7 @@ function sortData(event){
   } else if(event.target.id.includes("descending")){
     theData = sort(theData, selectSort.value, -1);
   }
-  printData();
+  printData(theData, cards, table, backBut, forwardBut, pageSelector, lines, inputToggle.checked);
 }
 
 function moveBetweenPages(event){
@@ -205,208 +292,9 @@ function moveBetweenPages(event){
     }
   }
   if (inputToggle.checked){
-    showTable(parseInt(pageSelector.value) + 1, theData);
+    showTable(theData, table, parseInt(pageSelector.value) + 1, lines)(parseInt(pageSelector.value) + 1, theData);
   } else{
-    showCards(parseInt(pageSelector.value) + 1, theData);
-  }
-}
-
-function createPaginator(totalPages){
-  while (pageSelector.firstChild) {
-    pageSelector.removeChild(pageSelector.firstChild);
-  }
-  for(let i = 0; i < totalPages; i++){
-    const option = document.createElement('option');
-    option.value = i;
-    option.text = i+1;
-    pageSelector.add(option);
-  }
-}
-
-function printData(){
-  totalPages = Math.ceil(theData.length/lines);
-  //REvisar que el alert salga una sola vez
-  if(totalPages === 0){
-    alert("Didn't find countries according to your searching parameters.");
-  }
-  if(totalPages === 1){
-    backBut.disabled = true;
-    forwardBut.disabled = true;
-  } else {
-    forwardBut.disabled = false;
-  }
-  createPaginator(totalPages);
-  if (inputToggle.checked){
-    showTable(parseInt(pageSelector.value) + 1, theData);
-  } else{
-    showCards(parseInt(pageSelector.value) + 1, theData);
-  }
-}
-
-function showTable(page, countries){
-  const initial = (page - 1)*lines;
-  const final = initial + lines;
-  if(countries.length !== 1){
-    const dataTable = countries.slice(initial, final);
-    createTable(page, dataTable);
-  } else{
-    createTable(page, countries);
-  }
-}
-
-function showCards(page, countries){
-  const initial = (page - 1)*lines;
-  const final = initial + lines;
-  if(countries.length !== 1){
-    const dataTable = countries.slice(initial, final);
-    createCards(page, dataTable);
-  } else{
-    createCards(page, countries);
-  }
-}
-
-function createTable(page, data){
-  while (table.firstChild) {
-    table.removeChild(table.firstChild);
-  }
-  const caption = document.createElement('caption');
-  caption.innerHTML = "Table of Countries";
-  table.append(caption);
-
-  //Fill the titles of the table
-  const thead = document.createElement('thead');
-  let tr = document.createElement('tr');
-  for (const i of titles){
-    const th = document.createElement('th');
-    th.innerHTML = i;
-    tr.append(th);
-  }
-  thead.append(tr);
-  table.append(thead);
-
-  // Fill the data of the table
-  /**Implementar esto con .map */
-  const tbody = document.createElement('tbody');
-  for (const i of data){
-    tr = document.createElement('tr');
-    for (const j of titles){
-      const td = document.createElement('td');
-      if (j === 'No'){
-        td.innerHTML = `${(page-1)*lines + (data.indexOf(i)+1)}`;
-      } else if (j === 'Country'){
-        let name = `${i.name.common}\t${i.flag}`;
-        (i.independent) ? name += '\t✅' : name += '\t❌';
-        //td.innerHTML = name;
-        const abbr = document.createElement('abbr');
-        abbr.title = i.name.official;
-        abbr.innerHTML = name;
-        td.appendChild(abbr);
-
-      } else if ( j === 'Capital'){
-        if (typeof(i.capital) === 'object'){
-          td.innerHTML = i.capital;
-        } else {
-          td.innerHTML = '❌';
-        }
-      } else if (j === 'Languages'){
-        let lang = "";
-        if(typeof(i.languages) === 'object'){
-          for(const key of Object.keys(i.languages)){
-            lang += `${key},\t`;
-          }
-        } else {
-          lang = "❌,\t";
-        }
-        td.innerHTML = lang.slice(0,-2);
-      } else if (j === 'Area'){
-        td.innerHTML = i.area;
-      } else if (j === 'Population'){
-        td.innerHTML = i.population;
-      } else if (j === 'Gini'){
-        if (typeof i.gini === 'object'){
-          for (const k of Object.keys(i.gini)){
-            //td.innerHTML = i.gini[`${k}`];
-            td;
-            const abbr = document.createElement('abbr');
-            abbr.title = k;
-            abbr.innerHTML = parseFloat(i.gini[`${k}`]).toFixed(1);
-            td.appendChild(abbr);
-          }
-        } else {
-          td.innerHTML = "❌";
-        }
-      }
-      tr.append(td);
-    }
-    tbody.append(tr);
-  }
-  table.append(tbody);
-}
-
-function createCards(page, countries){  
-  while (cards.firstChild) {
-    cards.removeChild(cards.firstChild);
-  }
-
-  for (const i of countries){
-    const div1 = document.createElement('div');
-    div1.style.margin = '10px';
-    div1.style.height = '250px';
-    div1.className = "flip-card";
-
-    const div2 = document.createElement('div');
-    div2.className = "flip-card-inner";
-
-    const div3 = document.createElement('div');
-    div3.className = "flip-card-front";
-    const img = document.createElement('img');
-    img.src = i.flags.png;
-    img.alt = i.flags.alt;
-    img.style.width = '100%';
-    img.style.height = '200px';
-    img.style.border = 'solid';
-    div3.style.alignSelf = "center";
-
-    const div4 = document.createElement('div');
-    div4.className = 'flip-card-back';
-    div4.style.border = 'solid';
-    const h4 = document.createElement('h1');
-    h4.innerHTML = `${i.name.common}`;
-    const h6 = document.createElement('h6');
-    h6.innerHTML = `${i.name.official}\t${i.independent ? '\t✅' : '\t❌'}`;
-    //h1.append(h3);
-    const p1 = document.createElement('p');
-    p1.innerHTML = (typeof i.capital === 'object') ? `Capital: ${i.capital[0]}`: `Capital: ❌`;
-    const p2 = document.createElement('p');
-    p2.innerHTML = (typeof i.area === 'number') ? `Area: ${i.area}`: `Area: ❌`;
-    const p3 = document.createElement('p');
-    p3.innerHTML = (typeof i.population === 'number') ? `Population: ${i.population}`: `Population: ❌`;
-    
-    if(i.continents[0] === 'America'){
-      div4.style.backgroundColor = '#FFFB7B';
-    } else if(i.continents[0] === 'Asia'){
-      div4.style.backgroundColor = '#CBADE0';
-    } else if(i.continents[0] === 'Europe'){
-      div4.style.backgroundColor = '#FCC2D2';
-    } else if(i.continents[0] === 'Africa'){
-      div4.style.backgroundColor = 'lightgreen';
-    } else if(i.continents[0] === 'Oceania'){
-      div4.style.backgroundColor = 'lightblue';
-    } else if(i.continents[0] === 'Antarctica'){
-      div4.style.backgroundColor = '#A3C7E3';
-    }
-    div4.style.color = "#1D0030";
-  
-    cards.append(div1);
-    div1.appendChild(div2);
-    div2.appendChild(div3);
-    div2.appendChild(div4);
-    div3.appendChild(img);
-    div4.appendChild(h4);
-    div4.append(h6);
-    div4.append(p1);
-    div4.append(p2);
-    div4.append(p3);
+    showCards(theData, cards, parseInt(pageSelector.value) + 1, lines)(parseInt(pageSelector.value) + 1, theData);
   }
 }
 
@@ -450,7 +338,8 @@ function createSubFilterOptions(){
 }
 
 function  createSelectForYears(){
-  const containerCanvasGini = document.querySelector('section[data-test="gini-canvas"]');
+  const containerCanvasGiniYear = document.querySelector('section[data-test="gini-canvas-year"]');
+  console.log();
   yearSelector = document.createElement('select');
   for (let i=0; i<globalData.length; i++){
     if('gini' in globalData[i]){
@@ -466,7 +355,12 @@ function  createSelectForYears(){
     option.text = i;
     yearSelector.append(option);
   }
-  //document.querySelector('#gini').append(yearSelector);
+  console.log(containerCanvasGiniYear.children.length);
+  console.log(containerCanvasGiniYear.children);
+  while (containerCanvasGiniYear.children[2]) {
+    containerCanvasGiniYear.removeChild(containerCanvasGiniYear.children[2]);
+  }
+  containerCanvasGiniYear.append(yearSelector);
 }
