+++ src/style.css
@@ -3,7 +3,7 @@
     font-size: 1vw;
     padding: 2vw;
     margin: 2vw;
-    border-radius: 0.3rem;
+    border-radius: 15px;
 }
 
 body{
@@ -120,6 +120,14 @@ article section div[id="pages"]{
 .flex-container[id="cards"]{
   flex-direction: row;
   flex-wrap: wrap;
+  padding: 1vw;
+}
+
+.flex-container[id="calculus-container"]{
+  flex-direction: row;
+  flex-wrap: nowrap;
+  align-items: start;
+  width: fit-content;
   padding: 1vw;
 }
 
@@ -300,4 +308,49 @@ input:checked + .slide:before {
 /* Add rounded corners to the top left and the top right corner of the image */
 img {
   border-radius: 5px 5px 0 0;
+}
+
+
+
+.pagination-container {
+  width: calc(100% - 2rem);
+  display: flex;
+  align-items: center;
+  position: absolute;
+  /*bottom: 0;*/
+  top: 10;
+  padding: 1rem 0;
+  justify-content: center;
+}
+
+.pagination-number,
+.pagination-button{
+  font-size: 1.1rem;
+  background-color: transparent;
+  border: none;
+  color: dodgerblue;
+  margin: 0.25rem 0.25rem;
+  cursor: pointer;
+  height: 2.5rem;
+  width: 2.5rem;
+  border-radius: .2rem;
+}
+
+.pagination-number:hover,
+.pagination-button:not(.disabled):hover {
+  background: #fff;
+  color: darkblue;
+}
+
+.pagination-number.active {
+  color: #fff;
+  background: #0085b6;
+}
+
+.calculus-container{
+  flex-direction: columns;
+}
+
+footer .heart {
+  color: #DC143C;
 }
